Modify docker-compose to work with new structure
A few variables had to be changed to fit the new structure. Nothing too
serious. The healthcheck section simply checks that the database is in
fact running correctly, as I've had an issue where the first time the
containers are run the web-app complains that the database isn't
running. I'll modify the readme to reflect this.